✨ 不要なページコンポーネントを削除し、ルーティングを更新
