--- src/pages/About.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import Card from "../components/Card";
-
-const About: React.FC = () => (
-  <Card className="max-w-md mx-auto my-8">
-    <h1 className="text-2xl font-bold mb-2">About Page</h1>
-    <p>This is a sample about page.</p>
-  </Card>
-);
-
-export default About;
