--- src/pages/Contact.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import Card from "../components/Card";
-
-const Contact: React.FC = () => (
-  <Card className="max-w-md mx-auto my-8">
-    <h1 className="text-2xl font-bold mb-2">Contact Page</h1>
-    <p>You can reach us at contact@example.com.</p>
-  </Card>
-);
-
-export default Contact;
