+++ src/pages/__Home.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import Card from "../components/Card";
+
+const Home: React.FC = () => (
+  <Card className="max-w-md mx-auto my-8">
+    <h1 className="text-2xl font-bold mb-2">Home Page</h1>
+    <p>Welcome to the home page!</p>
+  </Card>
+);
+
+export default Home;
